move handler code
