+++ packages/hono-takibi/src/generators/zod-openapi-hono/app/helper/get-route-maps.ts
@@ -1,5 +1,5 @@
 import type { OpenAPISpec } from '../../../../types'
-import { generateHandlerName } from '../../../handler/generate-handler-name'
+import { generateHandlerName } from '../../handler/generate-handler-name'
 import { generateRouteName } from '../../openapi/route/generate-route-name'
 
 export function getRouteMaps(openAPISpec: OpenAPISpec) {
