+++ packages/hono-takibi/src/generators/zod-openapi-hono/handler/generate-handler-name.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { generateHandlerName } from './generate-handler-name'
+import { generateHandlerName } from '../../handler/generate-handler-name'
 
 const generateHandlerNameTestCases: {
   method: string
