+++ packages/hono-takibi/src/generators/zod-openapi-hono/handler/generate-handler-name.ts
@@ -1,4 +1,4 @@
-import { generateRouteName } from '../zod-openapi-hono/openapi/route/generate-route-name'
+import { generateRouteName } from '../openapi/route/generate-route-name'
 
 export function generateHandlerName(method: string, path: string) {
   return `${generateRouteName(method, path)}Handler`
